Removed `preferDouble` rule (#1296)
This was discussed in more depth in issue #1139, where this rule in particular was not only causing several compiler errors but has also been proven to introduce subtle bugs into applications which may not be obvious or easy to diagnose at first look.

I think the value of this rule, even with it being opt-in, does not outweigh the risk to projects with it being turned on. As such, I believe this should be removed.

Alternative options may include a more prominent warning or description for the rule, or significant work on the rule itself to make it more resilient (but as this would essentially require knowledge of the source of every function called and whether that resides within the source code of the client) is rather impractical.